+++ src/components/travel-groups/[id]/cards/JoinRequestCard.tsx
